requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
astroid==2.15.2
black==23.3.0
cffi==1.15.1
click==8.1.3
coverage==7.2.3
cryptography==40.0.1
dill==0.3.6
Django==4.2
django-debug-toolbar==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy==1.2.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.6.0
pylint==2.17.2
pytz==2023.3
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.7
typing_extensions==4.5.0
wrapt==1.15.0